pyproject.toml
[project]
name = "embed_org_project"
version = "0.1.0"
description = "Platform that provides document management, generating embeddings, and vector storage capabilities, and endpoint usage for documents within your organization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.39.3",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-aws>=0.2.27",
    "langchain-community>=0.3.27",
    "langchain-postgres>=0.0.15",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "unstructured[docx,pdf]>=0.18.2",
    "uvicorn>=0.35.0",
]
